move exception handler tests
