+++ tests/unit/test_server_v1.py
@@ -296,27 +296,3 @@ class TestV1HttpApiOnStartup:
         assert app.state.manager_client is fake_client
         assert app.state.manager_platforms is fake_platforms
         assert app.state.settings is settings
-
-
-class TestServiceExceptionHandler:
-    def test_service_exception_handler_returns_correct_status_and_body(self) -> None:
-        """exception is converted to a Response with the right status."""
-        from ixmp4.server.v1 import V1HttpApi
-
-        exc = PlatformNotFound("test-platform")
-        request = mock.Mock()
-
-        response = V1HttpApi.service_exception_handler(request, exc)
-
-        assert response.status_code == exc.http_status_code
-
-    def test_service_exception_handler_works_for_forbidden(self) -> None:
-        """service_exception_handler handles any Ixmp4Error subclass."""
-        from ixmp4.server.v1 import V1HttpApi
-
-        exc = Forbidden("not allowed")
-        request = mock.Mock()
-
-        response = V1HttpApi.service_exception_handler(request, exc)
-
-        assert response.status_code == exc.http_status_code
